+++ nonebot-plugin-resolver/core/ytdlp.py
@@ -1,45 +1,42 @@
-import yt_dlp, asyncio
+import yt_dlp
+import random
+import asyncio
 
 from nonebot import logger
+from pathlib import Path
 
-async def get_video_title(url: str, is_oversea: bool, my_proxy=None, video_type='youtube') -> str:
+async def get_video_title(url: str, cookiefile: str = '', proxy: str = '') -> str:
     ydl_opts = {
         'quiet': True,
         'skip_download': True,
         'force_generic_extractor': True,
     }
-    if not is_oversea and my_proxy:
-        ydl_opts['proxy'] = my_proxy
-    if video_type == 'youtube':
-        ydl_opts['cookiefile'] = 'ytb_cookies.txt'
+    if proxy:
+        ydl_opts['proxy'] = proxy
+    if cookiefile:
+        ydl_opts['cookiefile'] = cookiefile
 
-    try:
-        with yt_dlp.YoutubeDL(ydl_opts) as ydl:
-            info_dict = await asyncio.to_thread(ydl.extract_info, url, download=False)
-            return info_dict.get('title', '-')
-    except Exception as e:
-        logger.error(f"Error: {e}")
-        return '-'
-
-async def download_ytb_video(url, is_oversea, path, my_proxy=None, video_type='youtube'):
+    with yt_dlp.YoutubeDL(ydl_opts) as ydl:
+        ydl.extract_info
+        info_dict = await asyncio.to_thread(ydl.extract_info, url, download=False)
+        return info_dict.get('title', '')
+        
+async def ytdlp_download_video(url: str, path: str, type: str, height: int = 1080, cookiefile: str = '', proxy: str = '') -> str:
+    filename = f"{path}/{type}-{random.randint(1, 10000)}"
     ydl_opts = {
-        'outtmpl': f'{path}/temp.%(ext)s',
+        'outtmpl': f'{filename}.%(ext)s',
         'merge_output_format': 'mp4',
     }
-    if video_type == 'youtube':
-        ydl_opts['cookiefile'] = 'ytb_cookies.txt'
-        if not 'shorts' in url:
-            ydl_opts['format'] = 'bv*[width=1280][height=720]+ba'
-    if not is_oversea and my_proxy:
-        ydl_opts['proxy'] = my_proxy
+    
+    if proxy:
+        ydl_opts['proxy'] = proxy
+    if cookiefile:
+        ydl_opts['cookiefile'] = cookiefile
 
-    try:
-        with yt_dlp.YoutubeDL(ydl_opts) as ydl:
-            await asyncio.to_thread(ydl.download, [url])
-        return f"{path}/temp.mp4"
-    except Exception as e:
-        print(f"Error: {e}")
-        return None
+    with yt_dlp.YoutubeDL(ydl_opts) as ydl:
+        await asyncio.to_thread(ydl.download, [url])
+    return f'{filename}.mp4'
+
