+++ pyproject.toml
@@ -19,6 +19,7 @@ yt-dlp = ">=2024.11.18"
 nonebot2 = ">=2.0.0-beta.5"
 nonebot-adapter-onebot = "^2.0.0-beta.1"
 nonebot_plugin_localstore = ">=0.6.0"
+nonebot-plugin-apscheduler = ">=0.5.0"
 
 [build-system]
 requires = ["poetry-core"]
